Experimental props styling API
This introduces a potential API for handling props
for gatsby-ui primitives and components in a theme.
This would be opt in, but could be automatically
parsed/analyzed and used to expose a higher-level
theming API for Gatsby themes.

The static functions, propTypes and propertyControls
can be used in tandem to expose what can and cannot
be modified. They can also be used to generate docs
which is a huge benefit. Allowing basic docs generation
for end users ensures that theming authors _start_
out from a basic docs foundation they can add onto.
Having an ecosystem of themes that are well documented
will allow a broader range of end users to use and
customize themes.

Here we'll go with the propertyControls API that
Framer X uses so that we can consumer Framer X components
without any additional work.

This can also be useful for exposing theme editing UI using
the same API. I think even more interestingly, we have the
potential to walk files in a given theme and _dynamically_
expose global props that can be edited.

For example, in this scenario we have the PostLink
propertyControls. We can statically analyze the component,
find where it's used and allow those props to be configured
without having to use component shadowing. Something similar
to this might be a desirable API for trivial changes to a
component. Changing the PostLink wrapper props to add in
{ px: 4 } should be able to be handled (IMO) without component
shadowing since that's a bit heavy-handed for trivial
prop modifications. It's a better fit for true overriding
of theme components and building more complex compositions.

Ref:
===

- #4
- https://github.com/jxnblk/styled-system/blob/master/src/index.js#L6-L16
- https://framer.gitbook.io/framer/components/code#exposing-controls
- https://github.com/johno/styled-api